ai-doc-backend/parsers_python/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=10.0.0
openpyxl==3.1.2
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
pymongo==4.6.1
python-multipart==0.0.6
PyPDF2==3.0.1
